refactor: 汎用的にCustomUIを作成できるようにベースコードは一か所に集約した
